fix(auth): correct signature calculation for OBS API compatibility
- Add newline after canonicalized headers when present in string-to-sign
- Sort query parameters alphabetically for consistent signature generation
- Add debug logging for StringToSign, CanonicalizedHeaders, and CanonicalizedResource
- Refactor object operations module into directory structure
- Add comprehensive integration tests covering all object operations
- Add tests for delete_objects with quiet mode and non-existent objects
